mix create and edit form to 1 component
